# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SnookerPocket API with Swagger",
    "description": "api description",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/speeldagen/{id}/speeldagVotes": {
      "put": {
        "tags": [
          "Speeldag"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Put new speeldagVote.",
            "schema": {
              "$ref": "#/definitions/PutSpeeldagvote"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/api/speeldagVotes/": {
      "get": {
        "tags": [
          "SpeeldagVote"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      }
    },
    "/api/speeldagVotes/{id}": {
      "get": {
        "tags": [
          "SpeeldagVote"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "SpeeldagVote"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "wedstrijdVotes": {
                  "example": "any"
                },
                "user": {
                  "example": "any"
                },
                "jokerGebruikt": {
                  "example": "any"
                },
                "SchiftingsvraagAntwoord": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/speeldagVotes/{speeldagId}/{userId}/votes": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "speeldagId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/speeldagVotes/update/{id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "jokerGebruikt": {
                  "example": "any"
                },
                "SchiftingsvraagAntwoord": {
                  "example": "any"
                },
                "wedstrijdVotes": {
                  "example": "any"
                },
                "user": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/wedstrijdVotes/": {
      "get": {
        "tags": [
          "WedstrijdVote"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": true
      },
      "post": {
        "tags": [
          "WedstrijdVote"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "vote": {
                  "example": "any"
                },
                "wedstrijd": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/api/wedstrijdVotes/{id}": {
      "get": {
        "tags": [
          "WedstrijdVote"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "deprecated": true
      }
    },
...
  },
  "definitions": {
...
    "PutSpeeldagvote": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "example": "ABC123"
        },
        "jokerGebruikt": {
          "type": "boolean",
          "example": false
        },
        "SchifingsvraagAntwoord": {
          "type": "number",
          "example": 12
        },
        "WedstrijdVotes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "wedstrijd": {
                "type": "string",
                "example": "1AB"
              },
              "vote": {
                "type": "string",
                "example": "1"
              }
            }
          }
        }
      }
    }
  }
}